+++ ts/store/reducers/types.ts
@@ -3,7 +3,7 @@ import { VersionInfoState } from "../../common/versionInfo/store/reducers/versio
 
 import { BonusState } from "../../features/bonus/bonusVacanze/store/reducers";
 import { PersistedFeaturesState } from "../../features/common/store/reducers";
-import { LollipopState } from "./lollipop";
+import { LollipopState } from "../../features/lollipop/store/reducers/lollipop";
 import { AppState } from "./appState";
 import { AssistanceToolsState } from "./assistanceTools";
 import { PersistedAuthenticationState } from "./authentication";
